# src/docs/blocktimestamp/swagger.yaml
definitions:
  integrated.Result:
    properties:
      account:
        type: string
      balance:
        type: string
      balanceHex:
        type: string
      tokenAddress:
        type: string
      transfer_history:
        items:
          $ref: '#/definitions/integrated.TransferHistory'
        type: array
    type: object
  integrated.TransferHistory:
    properties:
      from:
        type: string
      timestamp:
        type: integer
      to:
        type: string
      txHash:
        type: string
      value:
        type: string
      valueHex:
        type: string
    type: object
...
  router.ErrorResponse:
    properties:
      error:
        example: Description of the error
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /track/trackETH:
    get:
      description:
        Tracks Ethereum account activity, including transactions and balance
        changes, within a given date and target count.
      parameters:
        - description: Ethereum account address (e.g., '0x1234567890abcdef1234567890abcdef12345678')
          in: query
          name: account
          required: true
          type: string
        - description: Date in 'YYYY-MM-DD' format
          in: query
          name: date
          required: true
          type: string
        - description: Number of transactions to retrieve
          in: query
          name: targetCount
          type: integer
        - description: Time limit for processing in seconds
          in: query
          name: timeLimit
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/integrated.Result'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Track Ethereum account balance
      tags:
        - Track
swagger: '2.0'